Fix -Wshorten-64-to-32 warnings
These are sometimes useful (pointing out a couple clear issues in
the selftests), and mirror a warning that is enabled by default in
MSVC in our configuration.

This warning isn't enabled yet, but we may want to enable it.

+++ src/checkasm.c
@@ -644,7 +644,7 @@ static int set_cpu_affinity(const uint64_t affinity)
     BOOL(WINAPI * spdcs)(HANDLE, const ULONG *, ULONG) = (void *) GetProcAddress(
         GetModuleHandleW(L"kernel32.dll"), "SetProcessDefaultCpuSets");
     if (spdcs)
-        affinity_err = !spdcs(process, (ULONG[]) { affinity + 256 }, 1);
+        affinity_err = !spdcs(process, (ULONG[]) { (ULONG)affinity + 256 }, 1);
     else
   #endif
     {
@@ -1038,7 +1038,7 @@ static int parseu(unsigned *const dst, const char *const str, const int base)
     val   = strtoul(str, &end, base);
     if (errno || end == str || *end || val > (unsigned) -1)
         return 0;
-    *dst = val;
+    *dst = (unsigned) val;
     return 1;
 }
